refactor(Chats): Let drag and drop working same way in Windows and MacOS (WIP)
